+++ platform/Vulkan/VulkanBuffer.h
@@ -16,8 +16,8 @@ namespace exage::Graphics
         VulkanBuffer(VulkanBuffer&& old) noexcept;
         auto operator=(VulkanBuffer&& old) noexcept -> VulkanBuffer&;
 
-        void write(std::span<const std::byte> data, uint64_t offset) noexcept override;
-        void read(std::span<std::byte> data, uint64_t offset) const noexcept override;
+        void write(std::span<const std::byte> data, size_t offset) noexcept override;
+        void read(std::span<std::byte> data, size_t offset) const noexcept override;
 
         [[nodiscard]] auto getBuffer() const noexcept -> vk::Buffer { return _buffer; }
         [[nodiscard]] auto getAllocation() const noexcept -> vma::Allocation { return _allocation; }
